refactor : Popup > Modal로 네이밍 통일
